Configuration

+++ src/main/java/org/cos30018/hets/negotiation/strategy/TimeDependentStrategy.java
@@ -38,7 +38,7 @@ public class TimeDependentStrategy extends Strategy {
 		return Strategy.STRATEGY_TIME_DEPENDENT;
 	}
 
-	public double getDeadline() {
+	public int getDeadline() {
 		return deadline;
 	}
